add beast config command
