+++ cmd/beast/init.go
@@ -4,10 +4,8 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"github.com/BurntSushi/toml"
 	_ "github.com/lib/pq"
 	"github.com/sdslabs/beastv4/core"
-	"github.com/sdslabs/beastv4/core/config"
 	"github.com/sdslabs/beastv4/utils"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -24,14 +22,6 @@ const (
 	RESET       string = "\u001B[0m"
 	BLINK_ON    string = "\u001B[5m"
 	BLINK_OFF   string = "\u001B[25m"
-	DEFAULT_JWT string = "beast_jwt_secret_SUPER_STRONG_0x100010000100"
-)
-
-var (
-	AUTHORIZED_KEYS_FILE = filepath.Join(core.BEAST_GLOBAL_DIR, core.DEFAULT_AUTH_KEYS_FILE)
-
-	BEAST_GLOBAL_CONFIG  = filepath.Join(core.BEAST_GLOBAL_DIR, core.BEAST_CONFIG_FILE_NAME)
-	BEAST_EXAMPLE_CONFIG = filepath.Join(core.BEAST_GLOBAL_DIR, core.BEAST_EXAMPLE_DIR, core.BEAST_EX_CONFIG_FILE_NAME)
 )
 
 func initDirectories() error {
@@ -54,140 +44,6 @@ func initDirectories() error {
 		}
 	}
 
-	return nil
-}
-
-func initAuthorizedKeysFile() error {
-	log.Infoln("Defaulting Authorized keys file:", AUTHORIZED_KEYS_FILE, "... can be changed later")
-	return os.WriteFile(AUTHORIZED_KEYS_FILE, []byte("auth_keys"), 0666)
-}
-
-func downloadExampleBeastConfig() error {
-	response, err := http.Get("https://raw.githubusercontent.com/sdslabs/beast/master/_examples/example.config.toml")
-	if err != nil {
-		return err
-	}
-	defer response.Body.Close()
-
-	if response.StatusCode != http.StatusOK {
-		return errors.New("error while downloading: " + response.Status)
-	}
-
-	exampleConfig, err := os.Create(BEAST_EXAMPLE_CONFIG)
-	if err != nil {
-		return err
-	}
-	defer exampleConfig.Close()
-
-	_, err = io.Copy(exampleConfig, response.Body)
-	return err
-}
-
-func promptServerDetails(configuration *config.BeastConfig) {
-	for utils.PromptBinary("Configure an available server?") {
-		var server config.AvailableServer
-
-		server.Host = utils.PromptString("Enter Host Name, leave empty for localhost")
-		server.Username = utils.PromptString("Enter Username")
-		server.SSHKeyPath = utils.PromptString("Enter SSH Key Path")
-		server.Active = utils.PromptBinary("Enable this server?")
-
-		configuration.AvailableServers[server.Username] = server
-	}
-}
-
-func promptResourceLimits(configuration *config.BeastConfig) {
-	configuration.Memory = utils.PromptInt64("Default CPU Share:", 1024)
-	configuration.PidsLimit = utils.PromptInt64("Default PIDs Limit:", 100)
-	configuration.CPUShares = utils.PromptInt64("Default Memory Limit:", 1024)
-}
-
-func promptRemoteRepository(configuration *config.BeastConfig) {
-	for utils.PromptBinary("Configure a Remote Repository?") {
-		var remote config.GitRemote
-
-		remote.Url = utils.PromptString("Remote Repository URL, must be SSH based")
-		remote.Active = utils.PromptBinary("Enable this repository?")
-		remote.RemoteName = utils.PromptString("Remote Repository Name")
-		remote.Branch = utils.PromptString("Remote Repository Branch")
-		remote.Secret = utils.PromptSecret("Remote Repository SSH Key")
-
-		err := remote.ValidateGitConfig()
-		if err != nil {
-			log.Errorln(err)
-			log.Errorln("Skipping further repository configurations")
-			break
-		}
-
-		configuration.GitRemotes = append(configuration.GitRemotes, remote)
-	}
-}
-
-func promptCompetitionDetails(configuration *config.BeastConfig) {
-	configuration.CompetitionInfo.Name = utils.PromptString("Enter Competition Name")
-	configuration.CompetitionInfo.About = utils.PromptString("Enter Competition About Text")
-	configuration.CompetitionInfo.Prizes = utils.PromptString("Enter Competition Prizes Text")
-	configuration.CompetitionInfo.StartingTime = utils.PromptString("Enter Competition Start Time Text")
-	configuration.CompetitionInfo.EndingTime = utils.PromptString("Enter Competition End Time Text")
-	configuration.CompetitionInfo.LogoURL = utils.PromptString("Enter Competition Logo URL")
-	configuration.CompetitionInfo.DynamicScore = utils.PromptBinary("Enable Dynamic Scoring")
-}
-
-func promptNotificationWebhooks(configuration *config.BeastConfig) {
-	for utils.PromptBinary("Configure a Notification Webhook?") {
-		var notification config.NotificationWebhook
-
-		notification.ServiceName = utils.PromptSelection("Notification Service", core.NOTIFCIATION_SERVICES)
-		notification.URL = utils.PromptString("Notification Service URL")
-		notification.Active = utils.PromptBinary("Enable this webhook?")
-
-		configuration.NotificationWebhooks = append(configuration.NotificationWebhooks, notification)
-	}
-}
-
-func promptBeastConfiguration(configuration *config.BeastConfig) {
-	promptServerDetails(configuration)
-	promptResourceLimits(configuration)
-	promptRemoteRepository(configuration)
-	promptCompetitionDetails(configuration)
-	promptNotificationWebhooks(configuration)
-}
-
-func tryCopyExampleConfig() error {
-	var configuration config.BeastConfig
-
-	if _, err := os.Stat(BEAST_EXAMPLE_CONFIG); os.IsNotExist(err) {
-		if err = downloadExampleBeastConfig(); err != nil {
-			return err
-		}
-	}
-
-	if _, err := toml.Decode(BEAST_EXAMPLE_CONFIG, &configuration); err != nil {
-		return err
-	}
-
-	promptBeastConfiguration(&configuration)
-
-	file, err := os.Create(BEAST_GLOBAL_CONFIG)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	encoder := toml.NewEncoder(file)
-	if err = encoder.Encode(configuration); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func initBeastConfig() error {
-	if _, err := os.Stat(BEAST_GLOBAL_CONFIG); os.IsNotExist(err) {
-		return tryCopyExampleConfig()
-	}
-
-	log.Infoln("Found global config file:", BEAST_GLOBAL_CONFIG)
 	return nil
 }
 
@@ -320,7 +176,7 @@ func runBeastBootsteps() error {
 		return err
 	}
 
-	log.Infoln(fmt.Sprintf("Created %s", BEAST_GLOBAL_CONFIG))
+	log.Infoln(fmt.Sprintf("Beast global config file initiliased at %s", BEAST_GLOBAL_CONFIG))
 
 	if err := checkDockerDaemon(); err != nil {
 		return err
